Rename and simplify example data. Add focal clades.
Sorry to mix these in a single commit.

Focal clades are a new feature to highlight clades in plots and ease the making of subtrees.
For the fun I implemented them in S7…

Lazy-loaded example data has very generic names and might be polluting the namespace. I also
took opportunity to change identity to difference and invert the scrambling scale.

Both changes became convoluted through function examples and the update of the main vignette.